Move flask app parts into blueprints
plus related renaming and import style cleanups
