+++ fuzzbucket/blueprints/__init__.py
@@ -1,0 +1,3 @@
+from . import boxes, guts, image_aliases, keys
+
+__all__: tuple[str, ...] = ("boxes", "guts", "image_aliases", "keys")
