+++ tests/test_fuzzbucket.py
@@ -1,141 +1,14 @@
-import base64
-import collections
-import datetime
-import decimal
-import json
-import os
-import random
-import re
 import typing
 
-import boto3
-import botocore.exceptions
-import pytest
-from flask import Response
-from moto import mock_dynamodb, mock_ec2
-from werkzeug.exceptions import InternalServerError
-
 import fuzzbucket
-import fuzzbucket.app
-import fuzzbucket.cfg as cfg
-import fuzzbucket.datetime_ext
-import fuzzbucket.flask_dance_storage
-import fuzzbucket.reaper
-from fuzzbucket.app import app
-from fuzzbucket.box import Box
-
-TemplateResponse = collections.namedtuple(
-    "TemplateResponse", ("template_name", "status_code")
-)
-FakeRequest = collections.namedtuple(
-    "FakeRequest", ("headers", "args", "body", "environ")
-)
-WrappedError = collections.namedtuple("WrappedError", ("original_exception",))
-
-AnyDict = dict[str, typing.Any]
+from fuzzbucket import auth, blueprints, reaper
 
 
-@pytest.fixture(autouse=True)
-def resetti():
-    os.environ.setdefault("FUZZBUCKET_DEFAULT_VPC", "vpc-fafafafaf")
-    os.environ.setdefault("FUZZBUCKET_STAGE", "test")
-    fuzzbucket.get_dynamodb.cache_clear()
-    fuzzbucket.get_ec2_client.cache_clear()
-    fuzzbucket.app.app.testing = True
-    fuzzbucket.app.app.secret_key = f":hushed:-:open_mouth:-{random.randint(42, 666)}"
-    session_storage = fuzzbucket.flask_dance_storage.FlaskDanceStorage(
-        table_name=f"fuzzbucket-{os.getenv('FUZZBUCKET_STAGE')}-users"
-    )
-    fuzzbucket.app.app.config["session_storage"] = session_storage
-    fuzzbucket.app.app.config["oauth_blueprint"].storage = session_storage
+def test_deferred_app(fake_oauth_session, monkeypatch):
+    fake_oauth_session.authorized = False
+    monkeypatch.setattr(auth, "github", fake_oauth_session)
+    monkeypatch.setattr(blueprints.guts, "github", fake_oauth_session)
 
-
-@pytest.fixture
-def authd_headers() -> typing.List[typing.Tuple[str, str]]:
-    return [("Fuzzbucket-User", "pytest"), ("Fuzzbucket-Secret", "zzz")]
-
-
-@pytest.fixture
-def fake_oauth_session():
-    class FakeOAuthSession:
-        def __init__(self):
-            self.authorized = True
-            self.responses = {
-                "/user": {"login": "pytest"},
-                "/user/orgs": [{"login": "frob"}],
-                "userinfo": {"email": "pytest@example.org"},
-            }
-            self.next_json = None
-            self._token = None
-
-        @property
-        def token(self):
-            return self._token
-
-        @token.setter
-        def token(self, value):
-            self._token = value
-
-        @token.deleter
-        def token(self):
-            self._token = None
-
-        def get(self, path):
-            self.next_json = self.responses.get(path)
-            return self
-
-        def json(self):
-            response = self.next_json
-            self.next_json = None
-            return response
-
-    return FakeOAuthSession()
-
-
-@pytest.fixture
-def pubkey() -> str:
-    return "".join(
-        [
-            "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCcKKyTEzdI6zFMEmhbXSLemjTskw620yumv",
-            "bhoGwrY4zun/1cz+obxk1DZ+j0AfVTA9EQCr7AsFX3KRrevEBgHvWcK3vDp2h2pz/naM40SwF",
-            "dLK1+2G8vFy6zWZlFvQSNj8D6pxKGb6e0I3oVRBPd1V8z0AIswe2/9BiDi1K3Mx4yDoidZwnU",
-            "qweCCWwv3Y6nHkveEtVZlm8btGrlo2ya4IdCV2/KUK7FDbhGkLS7ZidVi+hS2GcrOTZYAkQW5",
-            "aS6r/QYTQGz94RjmyOFam5GhW5zboFdYnF9QD4WUGr4Gn9iI6QxaV50UXv37v+6pCaNYMPUjI",
-            f"SFQFMNhHnnMwcnx pytest@nowhere{random.randint(100, 999)}",
-        ]
-    )
-
-
-def setup_dynamodb_tables(dynamodb):
-    image_aliases_table = f"fuzzbucket-{os.getenv('FUZZBUCKET_STAGE')}-image-aliases"
-    table = dynamodb.create_table(
-        AttributeDefinitions=[dict(AttributeName="alias", AttributeType="S")],
-        KeySchema=[dict(AttributeName="alias", KeyType="HASH")],
-        TableName=image_aliases_table,
-        BillingMode="PAY_PER_REQUEST",
-    )
-    table.meta.client.get_waiter("table_exists").wait(TableName=image_aliases_table)
-
-    for alias, ami in {
-        "ubuntu18": "ami-fafafafafaf",
-        "rhel8": "ami-fafafafafaa",
-    }.items():
-        table.put_item(Item=dict(user="pytest", alias=alias, ami=ami))
-
-    users_table = f"fuzzbucket-{os.getenv('FUZZBUCKET_STAGE')}-users"
-    table = dynamodb.create_table(
-        AttributeDefinitions=[dict(AttributeName="user", AttributeType="S")],
-        KeySchema=[dict(AttributeName="user", KeyType="HASH")],
-        TableName=users_table,
-        BillingMode="PAY_PER_REQUEST",
-    )
-    table.meta.client.get_waiter("table_exists").wait(TableName=users_table)
-
-    for user, secret in {"pytest": "zzz", "nerf": "herder"}.items():
-        table.put_item(Item=dict(user=user, secret=secret))
-
-
-def test_deferred_app():
     state = {}
 
     def fake_start_response(status, headers):
@@ -152,7 +25,7 @@ def test_deferred_app():
         fake_start_response,
     )
     assert response is not None
-    assert state["status"] == "405 METHOD NOT ALLOWED"
+    assert state["status"][0] == "4"
     assert dict(state["headers"])["Content-Length"] > "0"
 
 
@@ -162,1955 +35,7 @@ def test_deferred_reap_boxes(monkeypatch):
     def fake_reap_boxes(event, context):
         state.update(event=event, context=context)
 
-    monkeypatch.setattr(fuzzbucket.reaper, "reap_boxes", fake_reap_boxes)
+    monkeypatch.setattr(reaper, "reap_boxes", fake_reap_boxes)
     fuzzbucket.deferred_reap_boxes({"oh": "hai"}, {"pro": "image"})
     assert state["event"] == {"oh": "hai"}
     assert state["context"] == {"pro": "image"}
-
-
-def test_get_ec2_client(monkeypatch):
-    state = {}
-
-    def fake_client(name):
-        state.update(name=name)
-        return "client"
-
-    monkeypatch.setattr(boto3, "client", fake_client)
-    client = fuzzbucket.get_ec2_client()
-    assert state["name"] == "ec2"
-    assert client == "client"
-
-
-@pytest.mark.parametrize("offline", [True, False], ids=["offline", "online"])
-def test_get_dynamodb(monkeypatch, offline):
-    state = {}
-
-    def fake_resource(name, **kwargs):
-        state.update(name=name, kwargs=kwargs)
-        return "resource"
-
-    monkeypatch.setattr(boto3, "resource", fake_resource)
-    if offline:
-        monkeypatch.setenv("IS_OFFLINE", "yep")
-    elif "IS_OFFLINE" in os.environ:
-        monkeypatch.delenv("IS_OFFLINE")
-
-    fuzzbucket.get_dynamodb.cache_clear()
-    resource = fuzzbucket.get_dynamodb()
-    assert state["name"] == "dynamodb"
-    assert resource == "resource"
-    if offline:
-        assert state["kwargs"]["region_name"] == "localhost"
-        assert state["kwargs"]["endpoint_url"] == "http://localhost:8000"
-
-
-def test_json_encoder():
-    class WithAsJson:
-        def as_json(self):
-            return {"golden": "feelings"}
-
-    class Dictish:
-        def __init__(self):
-            self.mellow = "gold"
-
-    other = ("odelay", ["mut", "ati", "ons"])
-
-    def enc(thing):
-        return fuzzbucket.AsJSONProvider(app).dumps(thing)
-
-    assert enc(WithAsJson()) == '{"golden": "feelings"}'
-    assert enc(Dictish()) == '{"mellow": "gold"}'
-    assert enc(other) == '["odelay", ["mut", "ati", "ons"]]'
-
-
-@mock_ec2
-def test_list_vpc_boxes(monkeypatch):
-    state = {}
-
-    def fake_list_boxes_filtered(ec2_client, filters):
-        state.update(ec2_client=ec2_client, filters=filters)
-        return ["ok"]
-
-    monkeypatch.setattr(fuzzbucket, "list_boxes_filtered", fake_list_boxes_filtered)
-
-    ec2_client = {"ec2_client": "sure"}
-    vpc_id = "vpc-fafafafaf"
-    listed = fuzzbucket.list_vpc_boxes(ec2_client, vpc_id)
-    assert listed == ["ok"]
-    assert state["ec2_client"] == ec2_client
-    for default_filter in [dict(f) for f in fuzzbucket.DEFAULT_FILTERS]:
-        assert default_filter in state["filters"]
-    assert {"Name": "vpc-id", "Values": [vpc_id]} in state["filters"]
-
-
-@pytest.mark.parametrize(
-    ("exc", "check_html", "err_match"),
-    [
-        pytest.param(
-            WrappedError(ValueError("not enough pylons")),
-            True,
-            "^NOPE=.*not enough pylons",
-            id="wrapped",
-        ),
-        pytest.param(
-            InternalServerError(
-                "spline reticulation overload",
-                response=Response(content_type="application/json"),
-            ),
-            False,
-            ".*spline reticulation overload",
-            id="wrapped_with_response",
-        ),
-        pytest.param(ValueError("whups"), True, "^NOPE.*whups", id="unhandled"),
-    ],
-)
-def test_handle_500(monkeypatch, exc, check_html, err_match):
-    state = {}
-
-    def fake_render_template(template_name, **kwargs):
-        state.update(template_name=template_name, kwargs=kwargs)
-        return "RENDERED"
-
-    monkeypatch.setattr(fuzzbucket.app, "render_template", fake_render_template)
-
-    if check_html:
-        assert fuzzbucket.app.handle_500(exc) == ("RENDERED", 500)
-        assert "kwargs" in state
-        assert "error" in state["kwargs"]
-        assert re.search(err_match, state["kwargs"]["error"]) is not None
-    else:
-        body, status = fuzzbucket.app.handle_500(exc)
-        assert status == 500
-        body_json = json.loads(body)
-        assert "error" in body_json
-        assert re.search(err_match, body_json["error"]) is not None
-
-
-@pytest.mark.parametrize(
-    (
-        "session_user",
-        "header_user",
-        "arg_user",
-        "authd",
-        "user_response",
-        "userinfo_response",
-        "expected_user",
-    ),
-    [
-        pytest.param(
-            "session-pytest",
-            None,
-            None,
-            True,
-            {"login": "session-pytest"},
-            {"email": "session-pytest"},
-            "session-pytest",
-            id="happy_session",
-        ),
-        pytest.param(
-            None,
-            "header-pytest",
-            None,
-            True,
-            {"login": "header-pytest"},
-            {"email": "header-pytest"},
-            "header-pytest",
-            id="happy_header",
-        ),
-        pytest.param(
-            None,
-            None,
-            "arg-pytest",
-            True,
-            {"login": "arg-pytest"},
-            {"email": "arg-pytest"},
-            "arg-pytest",
-            id="happy_arg",
-        ),
-        pytest.param(
-            None,
-            None,
-            None,
-            False,
-            {},
-            {},
-            None,
-            id="happy_nothing",
-        ),
-        pytest.param(
-            "session-pytest",
-            "session-pytest",
-            "session-pytest",
-            False,
-            {},
-            {},
-            None,
-            id="expired_token",
-        ),
-        pytest.param(
-            "elmer",
-            None,
-            None,
-            False,
-            {"login": "bugs"},
-            {"email": "bugs"},
-            None,
-            id="mismatched",
-        ),
-    ],
-)
-def test_set_user(
-    monkeypatch,
-    fake_oauth_session,
-    session_user,
-    header_user,
-    arg_user,
-    authd,
-    user_response,
-    userinfo_response,
-    expected_user,
-):
-    monkeypatch.setattr(fuzzbucket.app, "github", fake_oauth_session)
-    monkeypatch.setattr(fuzzbucket.app.oauth_blueprint, "session", fake_oauth_session)
-
-    fake_oauth_session.responses["/user"] = user_response
-    fake_oauth_session.responses["userinfo"] = userinfo_response
-    fake_oauth_session.authorized = authd
-    fake_oauth_session.token = "unchanged"
-
-    def fake_is_fully_authd():
-        return authd
-
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", fake_is_fully_authd)
-
-    fake_session = {"user": session_user}
-
-    monkeypatch.setattr(fuzzbucket.app, "session", fake_session)
-    monkeypatch.setattr(
-        fuzzbucket.app,
-        "request",
-        FakeRequest({"Fuzzbucket-User": header_user}, {"user": arg_user}, None, {}),
-    )
-
-    with app.test_client() as c:
-        response = c.get("/whoami")
-        assert response is not None
-        assert response.json is not None
-        assert response.json.get("you") == expected_user
-
-
-@pytest.mark.parametrize(
-    (
-        "authd",
-        "session_user",
-        "user_response",
-        "userinfo_response",
-        "secret_header",
-        "db_secret",
-        "expected",
-    ),
-    [
-        pytest.param(
-            True,
-            "pytest",
-            {"login": "pytest"},
-            {"email": "pytest@example.org"},
-            "verysecret",
-            "verysecret",
-            True,
-            id="happy",
-        ),
-        pytest.param(
-            True,
-            "pytest",
-            {"login": "pytest"},
-            {"email": "pytest@example.org"},
-            "verysecret",
-            "oh no",
-            False,
-            id="mismatched_db_secret",
-        ),
-        pytest.param(
-            True,
-            "pytest",
-            {"login": "pytest"},
-            {"email": "pytest@example.org"},
-            "well then",
-            "verysecret",
-            False,
-            id="mismatched_header_secret",
-        ),
-        pytest.param(
-            False,
-            "pytest",
-            {"login": "pytest"},
-            {"email": "pytest@example.org"},
-            "verysecret",
-            "verysecret",
-            False,
-            id="unauthorized",
-        ),
-    ]
-    + (
-        [
-            pytest.param(
-                True,
-                "pytest",
-                {"login": "founderman"},
-                {"email": "pytest@example.org"},
-                "verysecret",
-                "verysecret",
-                False,
-                id="mismatched_user",
-            ),
-            pytest.param(
-                True,
-                "superperson",
-                {"login": "pytest"},
-                {"email": "pytest@example.org"},
-                "verysecret",
-                "verysecret",
-                False,
-                id="mismatched_session_user",
-            ),
-        ]
-        if fuzzbucket.app.AUTH_PROVIDER == "github-oauth"
-        else []
-    ),
-)
-@mock_dynamodb
-def test_is_fully_authd(
-    monkeypatch,
-    fake_oauth_session,
-    authd,
-    session_user,
-    user_response,
-    userinfo_response,
-    secret_header,
-    db_secret,
-    expected,
-):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-
-    fake_oauth_session.responses["/user"] = user_response
-    fake_oauth_session.responses["userinfo"] = userinfo_response
-    fake_oauth_session.authorized = authd
-
-    if fuzzbucket.app.AUTH_PROVIDER == "github-oauth":
-        monkeypatch.setattr(fuzzbucket.app, "github", fake_oauth_session)
-
-    elif fuzzbucket.app.AUTH_PROVIDER == "oauth":
-        monkeypatch.setattr(
-            fuzzbucket.app.oauth_blueprint, "session", fake_oauth_session
-        )
-
-    monkeypatch.setattr(fuzzbucket.app, "session", {"user": session_user})
-    monkeypatch.setattr(
-        fuzzbucket.flask_dance_storage, "session", {"user": session_user}
-    )
-    monkeypatch.setattr(
-        fuzzbucket.app,
-        "request",
-        FakeRequest({"Fuzzbucket-Secret": secret_header}, {}, None, {}),
-    )
-
-    def fake_dump(*_):
-        return dict(user=session_user, secret=db_secret, token=None)
-
-    monkeypatch.setattr(app.config["session_storage"], "dump", fake_dump)
-
-    with app.app_context():
-        assert fuzzbucket.app.is_fully_authd() == expected
-
-
-def test_login(monkeypatch):
-    session = {}
-    monkeypatch.setattr(fuzzbucket.app, "session", session)
-
-    response = None
-    with app.test_client() as c:
-        response = c.get("/_login?user=pytest")
-
-    assert response is not None
-    assert response.status_code == 302
-    assert "Location" in response.headers
-    assert "user" in session
-    assert session["user"] == "pytest"
-
-
-@pytest.mark.parametrize(
-    ("authd", "session_user", "expected_status"),
-    [
-        pytest.param(True, "pytest", 204, id="happy"),
-        pytest.param(False, "pytest", 400, id="not_logged_in"),
-        pytest.param(True, "eggs", 404, id="unknown_user"),
-    ],
-)
-@mock_dynamodb
-def test_logout(monkeypatch, authd, session_user, expected_status):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-
-    session = {"user": session_user}
-    monkeypatch.setattr(fuzzbucket.app, "session", session)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    response = None
-    with app.test_client() as c:
-        response = c.post("/_logout")
-
-    assert response is not None
-    assert response.status_code == expected_status
-
-    if expected_status == 204:
-        assert "user" not in session
-
-
-@pytest.mark.parametrize(
-    ("allowed_orgs", "orgs_response", "raises", "expected"),
-    [
-        pytest.param(
-            {"frobs", "globs"},
-            [{"login": "frobs"}],
-            False,
-            TemplateResponse("auth_complete.html", 200),
-            id="happy",
-        ),
-        pytest.param(
-            {"frobs", "globs"},
-            [],
-            False,
-            TemplateResponse("error.html", 403),
-            id="forbidden",
-        ),
-        pytest.param(
-            {"frobs", "globs"},
-            {"message": "now you have gone and also done it"},
-            False,
-            TemplateResponse("error.html", 503),
-            id="github_err",
-        ),
-        pytest.param(
-            {"blubs", "glubs"},
-            [{"login": "blubs"}],
-            True,
-            TemplateResponse("error.html", 500),
-            id="no_secret_err",
-        ),
-    ],
-)
-@mock_dynamodb
-def test_auth_complete(
-    monkeypatch, fake_oauth_session, allowed_orgs, orgs_response, raises, expected
-):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-
-    state = {}
-
-    def fake_render_template(template_name, **kwargs):
-        state.update(template_name=template_name, kwargs=kwargs)
-        return "RENDERED"
-
-    monkeypatch.setattr(fuzzbucket.app, "render_template", fake_render_template)
-    monkeypatch.setattr(fuzzbucket.app, "github", fake_oauth_session)
-    monkeypatch.setattr(fuzzbucket.app, "session", {"user": "pytest"})
-    monkeypatch.setattr(fuzzbucket.flask_dance_storage, "session", {"user": "pytest"})
-    fake_oauth_session.responses["/user/orgs"] = orgs_response
-    monkeypatch.setattr(fuzzbucket.app, "ALLOWED_GITHUB_ORGS", allowed_orgs)
-
-    with app.test_client() as c:
-        if raises:
-            fuzzbucket.app.session["user"] = None
-            fuzzbucket.flask_dance_storage.session["user"] = None
-
-        response = c.get("/auth-complete")
-
-        assert response is not None
-        assert not response.is_json
-        assert response.status_code == expected.status_code
-        assert state["template_name"] == expected.template_name
-
-
-@pytest.mark.parametrize(
-    ("authd", "expected"),
-    [
-        pytest.param(
-            True,
-            200,
-            id="happy",
-        ),
-        pytest.param(False, 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_list_boxes(authd_headers, monkeypatch, authd, expected):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: boto3.client("ec2"))
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    response = None
-    with app.test_client() as c:
-        response = c.get("/", headers=authd_headers)
-    assert response is not None
-    assert response.status_code == expected
-    if authd:
-        assert response.json is not None
-        assert "boxes" in response.json
-        assert response.json["boxes"] is not None
-
-
-@pytest.mark.parametrize(
-    ("authd", "payload", "expected"),
-    [
-        pytest.param(
-            True,
-            dict(ami="ami-fafafafafaf"),
-            201,
-            id="happy",
-        ),
-        pytest.param(False, dict(ami="ami-fafafafafaf"), 403, id="forbidden"),
-        pytest.param(True, dict(ami="ami-ohno"), 400, id="bogus_ami"),
-        pytest.param(
-            True,
-            dict(ami="ami-fafafafafaf", root_volume_size=11),
-            201,
-            id="with_root_volume_size",
-        ),
-        pytest.param(
-            True,
-            dict(ami="ami-fafafafafaf", instance_tags={"fb:environment": "production"}),
-            201,
-            id="with_instance_tags",
-        ),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_create_box(authd_headers, monkeypatch, pubkey, authd, payload, expected):
-    ec2_client = boto3.client("ec2")
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(
-        fuzzbucket.flask_dance_storage, "get_dynamodb", lambda: dynamodb
-    )
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    if fuzzbucket.app.AUTH_PROVIDER == "oauth":
-        monkeypatch.setattr(fuzzbucket.app, "session", {"user": "lordtestingham"})
-
-        def fake_describe_key_pairs():
-            return {
-                "KeyPairs": [
-                    {
-                        "KeyName": "lordTestingham",
-                        "KeyPairId": "key-fafafafafafafafaf",
-                        "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                    }
-                ]
-            }
-
-        monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    def fake_describe_images(ImageIds=(), *_, **__):
-        return {
-            "ami-fafafafafaf": {
-                "Images": [
-                    {
-                        "RootDeviceName": "/dev/xyz",
-                        "BlockDeviceMappings": [
-                            {
-                                "DeviceName": "/dev/xyz",
-                                "Ebs": {"VolumeSize": 9},
-                            }
-                        ],
-                    }
-                ]
-            },
-        }.get(ImageIds[0], {"Images": []})
-
-    monkeypatch.setattr(ec2_client, "describe_images", fake_describe_images)
-
-    response = None
-    with monkeypatch.context() as mp:
-        mp.setattr(
-            fuzzbucket.app, "_fetch_first_compatible_github_key", lambda _: pubkey
-        )
-        with app.test_client() as c:
-            response = c.post(
-                "/box",
-                json=payload,
-                headers=authd_headers,
-            )
-    assert response is not None
-    assert response.status_code == expected
-    if authd and expected < 300:
-        assert response.json is not None
-        assert "boxes" in response.json
-        assert response.json["boxes"] != []
-
-
-@pytest.mark.parametrize(
-    ("authd", "update_body", "expected"),
-    [
-        pytest.param(
-            True,
-            {
-                "instance_tags": {"withered": "hand", "early": "seasons"},
-                "ttl": "108000.0",
-            },
-            200,
-            id="happy",
-        ),
-        pytest.param(
-            True,
-            {
-                "ttl": "108000.0",
-            },
-            200,
-            id="happy_ttl_only",
-        ),
-        pytest.param(
-            True,
-            {
-                "instance_tags": {"withered": "hand", "early": "seasons"},
-            },
-            200,
-            id="happy_tags_only",
-        ),
-        pytest.param(False, {}, 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_update_box(authd_headers, monkeypatch, pubkey, authd, update_body, expected):
-    ec2_client = boto3.client("ec2")
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(
-        fuzzbucket.flask_dance_storage, "get_dynamodb", lambda: dynamodb
-    )
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    def fake_describe_key_pairs():
-        return {
-            "KeyPairs": [
-                {
-                    "KeyName": "pytest",
-                    "KeyPairId": "key-fafafafafafafafaf",
-                    "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    def fake_describe_images(*_, **__):
-        return {
-            "Images": [
-                {
-                    "RootDeviceName": "/dev/xyz",
-                    "BlockDeviceMappings": [
-                        {"DeviceName": "/dev/xyz", "Ebs": {"VolumeSize": 9}}
-                    ],
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_images", fake_describe_images)
-
-    response = None
-    with monkeypatch.context() as mp:
-        mp.setattr(
-            fuzzbucket.app, "_fetch_first_compatible_github_key", lambda _: pubkey
-        )
-        with app.test_client() as c:
-            response = c.post(
-                "/box", json={"ami": "ami-fafafafafaf"}, headers=authd_headers
-            )
-    assert response is not None
-    assert "boxes" in typing.cast(AnyDict, response.json)
-
-    with app.test_client() as c:
-        with monkeypatch.context() as mp:
-            all_instances = ec2_client.describe_instances()
-
-            def fake_describe_instances(*_, **__):
-                return all_instances
-
-            mp.setattr(
-                ec2_client,
-                "describe_instances",
-                fake_describe_instances,
-            )
-            mp.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-            response = c.put(
-                f'/box/{typing.cast(AnyDict, response.json)["boxes"][0]["instance_id"]}',
-                json=update_body,
-                headers=authd_headers,
-            )
-            assert response.status_code == expected
-
-    if "ttl" in update_body:
-        re_fetched = fuzzbucket.list_boxes_filtered(ec2_client, [])
-        assert re_fetched[0].ttl == int(float(update_body["ttl"]))
-
-
-@pytest.mark.parametrize(
-    ("authd", "expected"),
-    [
-        pytest.param(
-            True,
-            204,
-            id="happy",
-        ),
-        pytest.param(False, 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_delete_box(authd_headers, monkeypatch, pubkey, authd, expected):
-    ec2_client = boto3.client("ec2")
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(
-        fuzzbucket.flask_dance_storage, "get_dynamodb", lambda: dynamodb
-    )
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    def fake_describe_key_pairs():
-        return {
-            "KeyPairs": [
-                {
-                    "KeyName": "pytest",
-                    "KeyPairId": "key-fafafafafafafafaf",
-                    "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    def fake_describe_images(*_, **__):
-        return {
-            "Images": [
-                {
-                    "RootDeviceName": "/dev/xyz",
-                    "BlockDeviceMappings": [
-                        {"DeviceName": "/dev/xyz", "Ebs": {"VolumeSize": 9}}
-                    ],
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_images", fake_describe_images)
-
-    response = None
-    with monkeypatch.context() as mp:
-        mp.setattr(
-            fuzzbucket.app, "_fetch_first_compatible_github_key", lambda _: pubkey
-        )
-        with app.test_client() as c:
-            response = c.post(
-                "/box", json={"ami": "ami-fafafafafaf"}, headers=authd_headers
-            )
-
-    assert response is not None
-    assert "boxes" in typing.cast(AnyDict, response.json)
-
-    with app.test_client() as c:
-        all_instances = ec2_client.describe_instances()
-
-        def fake_describe_instances(*_, **__):
-            return all_instances
-
-        monkeypatch.setattr(
-            ec2_client,
-            "describe_instances",
-            fake_describe_instances,
-        )
-        monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-        response = c.delete(
-            f'/box/{typing.cast(AnyDict, response.json)["boxes"][0]["instance_id"]}',
-            headers=authd_headers,
-        )
-        assert response.status_code == expected
-
-
-@mock_ec2
-@mock_dynamodb
-def test_delete_box_not_yours(monkeypatch, authd_headers, fake_oauth_session):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-
-    def fake_list_user_boxes(*_):
-        return []
-
-    monkeypatch.setattr(fuzzbucket.app, "list_user_boxes", fake_list_user_boxes)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-    monkeypatch.setattr(fuzzbucket.app, "github", fake_oauth_session)
-
-    response = None
-
-    with app.test_client() as c:
-        response = c.delete("/box/i-fafababacaca", headers=authd_headers)
-
-    assert response is not None
-    assert response.status_code == 403
-    assert "error" in typing.cast(AnyDict, response.json)
-    assert typing.cast(AnyDict, response.json)["error"] == "no touching"
-
-
-@pytest.mark.parametrize(
-    ("authd", "expected"),
-    [
-        pytest.param(
-            True,
-            204,
-            id="happy",
-        ),
-        pytest.param(False, 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_reboot_box(authd_headers, monkeypatch, pubkey, authd, expected):
-    ec2_client = boto3.client("ec2")
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(
-        fuzzbucket.flask_dance_storage, "get_dynamodb", lambda: dynamodb
-    )
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    def fake_describe_key_pairs():
-        return {
-            "KeyPairs": [
-                {
-                    "KeyName": "pytest",
-                    "KeyPairId": "key-fafafafafafafafaf",
-                    "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    def fake_describe_images(*_, **__):
-        return {
-            "Images": [
-                {
-                    "RootDeviceName": "/dev/xyz",
-                    "BlockDeviceMappings": [
-                        {"DeviceName": "/dev/xyz", "Ebs": {"VolumeSize": 9}}
-                    ],
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_images", fake_describe_images)
-
-    response = None
-    with monkeypatch.context() as mp:
-        mp.setattr(
-            fuzzbucket.app, "_fetch_first_compatible_github_key", lambda _: pubkey
-        )
-        with app.test_client() as c:
-            response = c.post(
-                "/box", json={"ami": "ami-fafafafafaf"}, headers=authd_headers
-            )
-
-    assert response is not None
-    assert "boxes" in typing.cast(AnyDict, response.json)
-
-    with app.test_client() as c:
-        all_instances = ec2_client.describe_instances()
-        with monkeypatch.context() as mp:
-
-            def fake_describe_instances(*_, **__):
-                return all_instances
-
-            mp.setattr(
-                ec2_client,
-                "describe_instances",
-                fake_describe_instances,
-            )
-            mp.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-            instance_id = typing.cast(AnyDict, response.json)["boxes"][0]["instance_id"]
-            response = c.post(
-                f"/reboot/{instance_id}",
-                headers=authd_headers,
-            )
-            assert response.status_code == expected
-
-
-@mock_ec2
-def test_reboot_box_not_yours(monkeypatch):
-    def fake_list_user_boxes(*_):
-        return []
-
-    monkeypatch.setattr(fuzzbucket.app, "list_user_boxes", fake_list_user_boxes)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    response = None
-    with app.test_client() as c:
-        response = c.post("/reboot/i-fafafafaf")
-    assert response is not None
-    assert response.status_code == 403
-    assert "error" in typing.cast(AnyDict, response.json)
-    assert typing.cast(AnyDict, response.json)["error"] == "no touching"
-
-
-@pytest.mark.parametrize(
-    ("authd", "expected"),
-    [
-        pytest.param(
-            True,
-            200,
-            id="happy",
-        ),
-        pytest.param(False, 403, id="forbidden"),
-    ],
-)
-@mock_dynamodb
-def test_list_image_aliases(authd_headers, monkeypatch, authd, expected):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    response = None
-    with app.test_client() as c:
-        response = c.get("/image-alias", headers=authd_headers)
-    assert response is not None
-    assert response.status_code == expected
-
-
-@pytest.mark.parametrize(
-    ("authd", "expected"),
-    [
-        pytest.param(
-            True,
-            201,
-            id="happy",
-        ),
-        pytest.param(False, 403, id="forbidden"),
-    ],
-)
-@mock_dynamodb
-def test_create_image_alias(authd_headers, monkeypatch, authd, expected):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    response = None
-    with app.test_client() as c:
-        response = c.post(
-            "/image-alias",
-            json={"alias": "yikes", "ami": "ami-fafababacaca"},
-            headers=authd_headers,
-        )
-    assert response is not None
-    assert response.status_code == expected
-
-
-@mock_dynamodb
-def test_create_image_alias_not_json(authd_headers, monkeypatch):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    response = None
-    with app.test_client() as c:
-        response = c.post(
-            "/image-alias",
-            data="HAY",
-            headers=authd_headers,
-        )
-    assert response is not None
-    assert response.status_code == 400
-
-
-@pytest.mark.parametrize(
-    ("authd", "expected"),
-    [pytest.param(True, 204, id="happy"), pytest.param(False, 403, id="forbidden")],
-)
-@mock_dynamodb
-def test_delete_image_alias(authd_headers, monkeypatch, authd, expected):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    response = None
-    with app.test_client() as c:
-        response = c.delete("/image-alias/ubuntu18", headers=authd_headers)
-    assert response is not None
-    assert response.status_code == expected
-
-
-@mock_dynamodb
-def test_delete_image_alias_no_alias(authd_headers, monkeypatch):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    response = None
-    with app.test_client() as c:
-        response = c.delete("/image-alias/nah", headers=authd_headers)
-    assert response is not None
-    assert response.status_code == 404
-
-
-@mock_dynamodb
-def test_delete_image_alias_not_yours(monkeypatch):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    response = None
-    with app.test_client() as c:
-        response = c.delete(
-            "/image-alias/ubuntu18",
-            headers=[("Authorization", base64.b64encode(b"jag:wagon").decode("utf-8"))],
-        )
-    assert response is not None
-    assert response.status_code == 403
-    assert "error" in typing.cast(AnyDict, response.json)
-    assert typing.cast(AnyDict, response.json)["error"] == "no touching"
-
-
-@pytest.mark.parametrize(
-    ("authd", "session_user", "key_alias", "expected"),
-    [
-        pytest.param(
-            True,
-            "lordtestingham",
-            "default",
-            200,
-            id="happy",
-        ),
-        pytest.param(
-            True,
-            "rumples",
-            "chuckit",
-            404,
-            id="missing",
-        ),
-        pytest.param(False, "foible", "default", 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_get_key(authd_headers, monkeypatch, authd, session_user, key_alias, expected):
-    ec2_client = boto3.client("ec2")
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    fake_session = {"user": session_user}
-    monkeypatch.setattr(fuzzbucket.app, "session", fake_session)
-
-    def fake_describe_key_pairs():
-        return {
-            "KeyPairs": [
-                {
-                    "KeyName": "lordTestingham",
-                    "KeyPairId": "key-fafafafafafafafaf",
-                    "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    request_path = "/key"
-    if key_alias != "default":
-        request_path = f"/key/{key_alias}"
-
-    response = None
-    with app.test_client() as c:
-        response = c.get(request_path, headers=authd_headers)
-    assert response is not None
-    assert response.status_code == expected
-    if authd and expected < 400:
-        assert response.json is not None
-        assert "key" in response.json
-        assert response.json["key"] is not None
-
-
-@pytest.mark.parametrize(
-    ("authd", "session_user", "n_keys", "expected"),
-    [
-        pytest.param(
-            True,
-            "lordtestingham",
-            2,
-            200,
-            id="happy",
-        ),
-        pytest.param(
-            True,
-            "rumples",
-            0,
-            200,
-            id="none",
-        ),
-        pytest.param(False, "foible", 0, 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_list_keys(authd_headers, monkeypatch, authd, session_user, n_keys, expected):
-    ec2_client = boto3.client("ec2")
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    fake_session = {"user": session_user}
-    monkeypatch.setattr(fuzzbucket.app, "session", fake_session)
-
-    def fake_describe_key_pairs():
-        return {
-            "KeyPairs": [
-                {
-                    "KeyName": "lordTestingham",
-                    "KeyPairId": "key-fafafafafafafafaf",
-                    "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                },
-                {
-                    "KeyName": "lordTestingham-fancy",
-                    "KeyPairId": "key-fafafafafafafafaf",
-                    "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                },
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    response = None
-    with app.test_client() as c:
-        response = c.get("/keys", headers=authd_headers)
-    assert response is not None
-    assert response.status_code == expected
-    if authd and expected < 400:
-        assert response.json is not None
-        assert "keys" in response.json
-        assert response.json["keys"] is not None
-        assert len(response.json["keys"]) == n_keys
-
-
-@pytest.mark.parametrize(
-    ("authd", "session_user", "key_alias", "request_kwargs", "expected"),
-    [
-        pytest.param(
-            True,
-            "philobuster",
-            "fancy",
-            dict(
-                json={
-                    "key_material": "".join(
-                        [
-                            "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCwZRcwdL1TLYM",
-                            "sKT6oYiKHjME0iyQKl1mOIZNA2pGqOJ8IH7UPX4AocNhw1G5xzA",
-                            "UG6FChZ32h8E+AMWjaJoOBnXSqlM3m1Up4KV0UsvPI5mVg/bm9j",
-                            "iCQ5OkwReEkmSC0hPAsQ5ztSZlRmG6Yo343D1wISgKOcmGEOdJR",
-                            "N26KiuIwSZ7LkMX1Uc1gIKaiNbTp8Jtn2nmB0O2R5Jvcsv5yICR",
-                            "jvTYl11hiNEg+TOJRQBoeyC2tsYwkWoabShm4Oi4X/UjB5UNDhG",
-                            "qQ/JX8XMyp0rFBIqTyd69csRoDqFJ2xGHYn+WmBbCHfyyks7LWz",
-                            "aCJzdekMg2iEBE7eoodM86V oop",
-                        ]
-                    )
-                }
-            ),
-            201,
-            id="happy",
-        ),
-        pytest.param(
-            True,
-            "slimer",
-            "default",
-            dict(
-                json={
-                    "key_material": "".join(
-                        [
-                            "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCwZRcwdL1TLYM",
-                            "sKT6oYiKHjME0iyQKl1mOIZNA2pGqOJ8IH7UPX4AocNhw1G5xzA",
-                            "UG6FChZ32h8E+AMWjaJoOBnXSqlM3m1Up4KV0UsvPI5mVg/bm9j",
-                            "iCQ5OkwReEkmSC0hPAsQ5ztSZlRmG6Yo343D1wISgKOcmGEOdJR",
-                            "N26KiuIwSZ7LkMX1Uc1gIKaiNbTp8Jtn2nmB0O2R5Jvcsv5yICR",
-                            "jvTYl11hiNEg+TOJRQBoeyC2tsYwkWoabShm4Oi4X/UjB5UNDhG",
-                            "qQ/JX8XMyp0rFBIqTyd69csRoDqFJ2xGHYn+WmBbCHfyyks7LWz",
-                            "aCJzdekMg2iEBE7eoodM86V oop",
-                        ]
-                    )
-                }
-            ),
-            500,
-            id="ec2_err",
-        ),
-        pytest.param(
-            True,
-            "philobuster",
-            "fancy",
-            dict(json={"key_material": "yuno"}),
-            400,
-            id="invalid_key_material",
-        ),
-        pytest.param(
-            True,
-            "philobuster",
-            "fancy",
-            dict(json={"key_material": ""}),
-            400,
-            id="empty_key_material",
-        ),
-        pytest.param(
-            True,
-            "philobuster",
-            "fancy",
-            dict(data="yuno"),
-            400,
-            id="not_json",
-        ),
-        pytest.param(
-            True,
-            "charizard",
-            "default",
-            dict(json={}),
-            409,
-            id="conflict",
-        ),
-        pytest.param(False, "morgenstern", "default", {}, 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_put_key(
-    authd_headers,
-    monkeypatch,
-    authd,
-    session_user,
-    key_alias,
-    request_kwargs,
-    expected,
-):
-    ec2_client = boto3.client("ec2")
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    fake_session = {"user": session_user}
-    monkeypatch.setattr(fuzzbucket.app, "session", fake_session)
-
-    state = {"describe_key_pairs_call": 0}
-
-    def fake_describe_key_pairs():
-        key_pairs = {
-            "0": {
-                "KeyPairs": [
-                    {
-                        "KeyName": "charizard",
-                        "KeyPairId": "key-fafafafafafafafaf",
-                        "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff"
-                        + ":aa:ff:aa:ff:aa",
-                    }
-                ]
-            },
-            "1": {
-                "KeyPairs": [
-                    {
-                        "KeyName": "charizard",
-                        "KeyPairId": "key-fafafafafafafafaf",
-                        "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff"
-                        + ":aa:ff:aa:ff:aa",
-                    },
-                    {
-                        "KeyName": "philobuster-fancy",
-                        "KeyPairId": "key-fafafafafafafafaf",
-                        "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff"
-                        + ":aa:ff:aa:ff:aa",
-                    },
-                ]
-            },
-        }[str(state["describe_key_pairs_call"])]
-        state["describe_key_pairs_call"] += 1
-        return key_pairs
-
-    monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    request_path = "/key"
-    if key_alias != "default":
-        request_path = f"/key/{key_alias}"
-
-    response = None
-    with app.test_client() as c:
-        response = c.put(request_path, headers=authd_headers, **request_kwargs)
-
-    assert response is not None
-    assert response.status_code == expected
-    if authd and expected < 400:
-        assert response.json is not None
-        assert "key" in response.json
-        assert response.json["key"] is not None
-
-
-@pytest.mark.parametrize(
-    ("authd", "session_user", "key_alias", "expected"),
-    [
-        pytest.param(
-            True,
-            "lordtestingham",
-            "default",
-            200,
-            id="happy",
-        ),
-        pytest.param(
-            True,
-            "rumples",
-            "chuckit",
-            404,
-            id="happy",
-        ),
-        pytest.param(False, "foible", "default", 403, id="forbidden"),
-    ],
-)
-@mock_ec2
-@mock_dynamodb
-def test_delete_key(
-    authd_headers, monkeypatch, authd, session_user, key_alias, expected
-):
-    ec2_client = boto3.client("ec2")
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: authd)
-
-    fake_session = {"user": session_user}
-    monkeypatch.setattr(fuzzbucket.app, "session", fake_session)
-
-    def fake_describe_key_pairs():
-        return {
-            "KeyPairs": [
-                {
-                    "KeyName": "lordTestingham",
-                    "KeyPairId": "key-fafafafafafafafaf",
-                    "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    request_path = "/key"
-    if key_alias != "default":
-        request_path = f"/key/{key_alias}"
-
-    response = None
-    with app.test_client() as c:
-        response = c.delete(request_path, headers=authd_headers)
-    assert response is not None
-    assert response.status_code == expected
-    if authd and expected < 400:
-        assert response.json is not None
-        assert "key" in response.json
-        assert response.json["key"] is not None
-
-
-@mock_dynamodb
-@pytest.mark.parametrize(
-    ("image_alias", "raises", "expected"),
-    [
-        pytest.param("noice", False, None, id="invalid"),
-        pytest.param("rhel8", True, None, id="errored"),
-        pytest.param("rhel8", False, "ami-fafafafafaa", id="valid"),
-    ],
-)
-def test_resolve_ami_alias(monkeypatch, image_alias, raises, expected):
-    monkeypatch.setenv("FUZZBUCKET_STAGE", "bogus")
-
-    dynamodb = boto3.resource("dynamodb")
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    setup_dynamodb_tables(dynamodb)
-
-    if raises:
-
-        def boom(*_):
-            raise botocore.exceptions.ClientError(
-                {"Error": {"Code": 1312, "Message": "nah"}}, "wut"  # type: ignore
-            )
-
-        monkeypatch.setattr(dynamodb, "Table", boom)
-
-    response = fuzzbucket.app._resolve_ami_alias(image_alias)
-    assert response == expected
-
-
-@pytest.mark.parametrize(
-    ("raises", "api_response", "expected_key"),
-    [
-        pytest.param(
-            False,
-            [
-                {"key": "ssh-ecdsagoop first"},
-                {"key": "ssh-rsa second"},
-                {"key": "ssh-rsa third"},
-            ],
-            "ssh-rsa second",
-            id="3_keys",
-        ),
-        pytest.param(
-            False,
-            [
-                {"key": "ssh-ecdsagoop first"},
-                {"key": "ssh-nope second"},
-                {"key": "ssh-ed25519 third"},
-            ],
-            "ssh-ed25519 third",
-            id="3_keys_also",
-        ),
-        pytest.param(False, [{"key": "ssh-rsa first"}], "ssh-rsa first", id="1_key"),
-        pytest.param(False, [], "", id="empty"),
-        pytest.param(
-            True,
-            [{"key": "first"}, {"key": "second"}, {"key": "third"}],
-            "",
-            id="err_3_keys",
-        ),
-        pytest.param(True, [{"key": "first"}], "", id="err_1_key"),
-        pytest.param(True, [], "", id="err_empty"),
-    ],
-)
-def test_fetch_first_compatible_github_key(
-    monkeypatch, raises, api_response, expected_key
-):
-    class FakeOAuthSession:
-        def get(self, *_):
-            if raises:
-                raise ValueError("oh no")
-            return self
-
-        def json(self):
-            return api_response
-
-    monkeypatch.setattr(fuzzbucket.app, "github", FakeOAuthSession())
-    assert fuzzbucket.app._fetch_first_compatible_github_key("user") == expected_key
-
-
-@mock_ec2
-@mock_dynamodb
-def test_reap_boxes(authd_headers, monkeypatch, pubkey):
-    ec2_client = boto3.client("ec2")
-    monkeypatch.setattr(fuzzbucket.app, "get_ec2_client", lambda: ec2_client)
-    monkeypatch.setattr(
-        fuzzbucket.reaper, "get_ec2_client", lambda: boto3.client("ec2")
-    )
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(fuzzbucket.app, "get_dynamodb", lambda: dynamodb)
-    monkeypatch.setattr(
-        fuzzbucket.flask_dance_storage, "get_dynamodb", lambda: dynamodb
-    )
-    monkeypatch.setattr(fuzzbucket.app, "is_fully_authd", lambda: True)
-
-    if fuzzbucket.app.AUTH_PROVIDER == "oauth":
-        monkeypatch.setattr(fuzzbucket.app, "session", {"user": "lordtestingham"})
-
-        def fake_describe_key_pairs():
-            return {
-                "KeyPairs": [
-                    {
-                        "KeyName": "lordTestingham",
-                        "KeyPairId": "key-fafafafafafafafaf",
-                        "KeyFingerprint": "ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa:ff:aa",
-                    }
-                ]
-            }
-
-        monkeypatch.setattr(ec2_client, "describe_key_pairs", fake_describe_key_pairs)
-
-    def fake_describe_images(*_, **__):
-        return {
-            "Images": [
-                {
-                    "RootDeviceName": "/dev/xyz",
-                    "BlockDeviceMappings": [
-                        {"DeviceName": "/dev/xyz", "Ebs": {"VolumeSize": 9}}
-                    ],
-                }
-            ]
-        }
-
-    monkeypatch.setattr(ec2_client, "describe_images", fake_describe_images)
-
-    response = None
-    with monkeypatch.context() as mp:
-        mp.setattr(
-            fuzzbucket.app, "_fetch_first_compatible_github_key", lambda _: pubkey
-        )
-        with app.test_client() as c:
-            response = c.post(
-                "/box",
-                json={"ttl": "-1", "ami": "ami-fafafafafaf"},
-                headers=authd_headers,
-            )
-    assert response is not None
-    assert "boxes" in typing.cast(AnyDict, response.json)
-    instance_id = typing.cast(AnyDict, response.json)["boxes"][0]["instance_id"]
-    assert instance_id != ""
-
-    the_future = fuzzbucket.utcnow() + datetime.timedelta(hours=1)
-
-    with monkeypatch.context() as mp:
-        ec2_client = boto3.client("ec2")
-
-        def fake_list_vpc_boxes(*_, **__):
-            ret = []
-            for box_dict in typing.cast(AnyDict, response.json)["boxes"]:
-                for virtual in ("age", "max_age"):
-                    if virtual in box_dict:
-                        box_dict.pop(virtual)
-                box = Box(**box_dict)
-                box.created_at = None
-                ret.append(box)
-            return ret
-
-        mp.setattr(fuzzbucket, "utcnow", lambda: the_future)
-        mp.setattr(fuzzbucket.reaper, "list_vpc_boxes", fake_list_vpc_boxes)
-        mp.setenv("FUZZBUCKET_DEFAULT_VPC", "vpc-fafafafafaf")
-        reap_response = fuzzbucket.reaper.reap_boxes(
-            None,
-            None,
-            ec2_client=ec2_client,
-        )
-        assert reap_response["reaped_instance_ids"] == []
-
-    with monkeypatch.context() as mp:
-        ec2_client = boto3.client("ec2")
-
-        def fake_list_vpc_boxes_2(*_, **__):
-            ret = []
-            for box_dict in typing.cast(AnyDict, response.json)["boxes"]:
-                for virtual in ("age", "max_age"):
-                    if virtual in box_dict:
-                        box_dict.pop(virtual)
-                box = Box(**box_dict)
-                box.ttl = None  # type: ignore
-                ret.append(box)
-            return ret
-
-        mp.setattr(fuzzbucket, "utcnow", lambda: the_future)
-        mp.setattr(fuzzbucket.reaper, "list_vpc_boxes", fake_list_vpc_boxes_2)
-        mp.setenv("FUZZBUCKET_DEFAULT_VPC", "vpc-fafafafafaf")
-        reap_response = fuzzbucket.reaper.reap_boxes(
-            None,
-            None,
-            ec2_client=ec2_client,
-        )
-        assert reap_response["reaped_instance_ids"] == []
-
-    with monkeypatch.context() as mp:
-        ec2_client = boto3.client("ec2")
-
-        def fake_list_vpc_boxes_3(*_, **__):
-            ret = []
-            for box_dict in typing.cast(AnyDict, response.json)["boxes"]:
-                for virtual in ("age", "max_age"):
-                    if virtual in box_dict:
-                        box_dict.pop(virtual)
-                box = Box(**box_dict)
-                ret.append(box)
-            return ret
-
-        mp.setattr(fuzzbucket, "utcnow", lambda: the_future)
-        mp.setattr(fuzzbucket.reaper, "list_vpc_boxes", fake_list_vpc_boxes_3)
-        mp.setenv("FUZZBUCKET_DEFAULT_VPC", "vpc-fafafafafaf")
-        reap_response = fuzzbucket.reaper.reap_boxes(
-            None,
-            None,
-            ec2_client=ec2_client,
-        )
-        assert reap_response["reaped_instance_ids"] != []
-
-    assert instance_id not in [
-        box.instance_id
-        for box in fuzzbucket.list_boxes_filtered(
-            ec2_client, [dict(f) for f in fuzzbucket.DEFAULT_FILTERS]
-        )
-    ]
-
-
-def test_box():
-    box = Box(instance_id="i-fafafafafafafaf")
-    assert box.age is None
-
-    box.created_at = str(
-        (fuzzbucket.utcnow() - datetime.timedelta(days=1, minutes=1)).timestamp()
-    )
-    assert box.age is not None
-    assert box.age.startswith("1 day,")
-
-    assert "instance_id" in box.as_json()
-    assert "age" in box.as_json()
-    assert "max_age" in box.as_json()
-
-    with pytest.raises(TypeError):
-        Box(instance_id="i-fafafafbabacaca", frobs=9001)  # type: ignore
-
-    dumped = app.json.dumps(box)
-    assert '"age":' in dumped
-    assert '"max_age":' in dumped
-
-
-@mock_dynamodb
-def test_flask_dance_storage(monkeypatch):
-    dynamodb = boto3.resource("dynamodb")
-    setup_dynamodb_tables(dynamodb)
-    monkeypatch.setattr(
-        fuzzbucket.flask_dance_storage, "get_dynamodb", lambda: dynamodb
-    )
-    fake_session = {"user": "pytest"}
-    monkeypatch.setattr(fuzzbucket.flask_dance_storage, "session", fake_session)
-
-    table_name = f"fuzzbucket-{os.getenv('FUZZBUCKET_STAGE')}-users"
-    storage = fuzzbucket.flask_dance_storage.FlaskDanceStorage(table_name)
-
-    storage.set(None, {"token": "jagwagon8999"})
-    storage.secret = "rotini"
-
-    actual = storage.dump()
-
-    for key, value in (
-        ("user", "pytest"),
-        ("token", {"token": "jagwagon8999"}),
-        ("secret", "rotini"),
-    ):
-        assert key in actual
-        assert actual[key] == value
-
-    storage.set(None, {"token": "jagwagon9000", "expires_at": decimal.Decimal(14)})
-    storage.secret = "farfalla"
-
-    actual = storage.dump()
-
-    for key, value in (
-        ("user", "pytest"),
-        ("token", {"token": "jagwagon9000", "expires_at": 14}),
-        ("secret", "farfalla"),
-    ):
-        assert key in actual
-        assert actual[key] == value
-
-    storage.delete(None)
-
-    assert storage.get(None) is None
-    assert storage.secret is None
-
-    with monkeypatch.context() as mp:
-        mp.setattr(
-            fuzzbucket.flask_dance_storage,
-            "session",
-            {"user": "Carbunkle"},
-        )
-
-        assert storage.dump() == {"user": "carbunkle"}
-
-        def fake_load_user():
-            return None
-
-        mp.setattr(storage, "_load_user", fake_load_user)
-
-        with pytest.raises(ValueError):
-            storage.set(None, "busytown321")
-
-        with pytest.raises(ValueError):
-            storage.secret = "crungus"
-
-        with pytest.raises(ValueError):
-            storage.delete(None)
-
-        assert storage.get(None) is None
-        assert storage.secret is None
-
-
-@pytest.mark.parametrize(
-    ("input_string", "expected", "expect_error"),
-    [
-        pytest.param(
-            "4:30:05",
-            datetime.timedelta(hours=4, minutes=30, seconds=5),
-            False,
-            id="only_sexagesimal",
-        ),
-        pytest.param(
-            "20:95",
-            datetime.timedelta(minutes=20, seconds=95),
-            False,
-            id="short_sexagesimal",
-        ),
-        pytest.param("4 days", datetime.timedelta(days=4), False, id="days"),
-        pytest.param("3 hours", datetime.timedelta(hours=3), False, id="hours"),
-        pytest.param("3 minutes", datetime.timedelta(minutes=3), False, id="minutes"),
-        pytest.param(
-            "900.5 seconds", datetime.timedelta(seconds=900.5), False, id="seconds"
-        ),
-        pytest.param("2 week", datetime.timedelta(weeks=2), False, id="weeks"),
-        pytest.param(
-            "1 days, 1:10:00",
-            datetime.timedelta(days=1, hours=1, minutes=10),
-            False,
-            id="stdlib_string",
-        ),
-        pytest.param(
-            "10800", datetime.timedelta(seconds=10800.0), False, id="seconds_only"
-        ),
-        pytest.param("parrots", None, True, id="bogus"),
-    ],
-)
-def test_parse_timedelta(input_string, expected, expect_error):
-    if expect_error:
-        with pytest.raises(ValueError):
-            fuzzbucket.datetime_ext.parse_timedelta(input_string)
-        return
-
-    assert expected == fuzzbucket.datetime_ext.parse_timedelta(input_string)
-
-
-@pytest.mark.parametrize(
-    ("env", "keys", "default", "expected"),
-    [
-        pytest.param(
-            {"OK_BUD": "giraffe"},
-            ("OK_BUD",),
-            None,
-            "giraffe",
-            id="simple",
-        ),
-        pytest.param(
-            {"OK_BUD": "giraffe"},
-            ("OH_NO_BUD",),
-            None,
-            None,
-            id="missing",
-        ),
-        pytest.param(
-            {"ALLIES": "bison", "OK_BUD": "giraffe"},
-            ("ALLIES", "OK_BUD"),
-            None,
-            "bison",
-            id="first_wins",
-        ),
-        pytest.param(
-            {"ALLIES": "", "OK_BUD": "giraffe"},
-            ("ALLIES", "OK_BUD"),
-            None,
-            "giraffe",
-            id="skip_empty_values",
-        ),
-        pytest.param(
-            {"ALLIES": "gibbon", "OK_BUD": "giraffe"},
-            ("", "", "", "OK_BUD"),
-            None,
-            "giraffe",
-            id="skip_empty_keys",
-        ),
-        pytest.param(
-            {"OK_BUD": "giraffe"},
-            ("", "", "", "OPE"),
-            8_001,
-            8_001,
-            id="default",
-        ),
-    ],
-)
-def test_cfg_get(
-    env: dict[str, str],
-    keys: tuple[str, ...],
-    default: str | None,
-    expected: str | None,
-    monkeypatch,
-):
-    for key, value in env.items():
-        monkeypatch.setenv(key, value)
-
-    assert cfg.get(*keys, default=default) == expected
-
-
-@pytest.mark.parametrize(
-    ("env", "keys", "default", "expected"),
-    [
-        pytest.param(
-            {"FIDDLES": "ok"},
-            ("FIDDLES",),
-            False,
-            True,
-            id="simple",
-        ),
-        pytest.param(
-            {"FIDDLES": "ok"},
-            ("FADDLES",),
-            False,
-            False,
-            id="missing",
-        ),
-        pytest.param(
-            {"FIDDLES": "ok", "FADDLES": "no"},
-            ("FIDDLES", "FADDLES"),
-            False,
-            True,
-            id="first_wins",
-        ),
-        pytest.param(
-            {"FIDDLES": "", "FADDLES": "no"},
-            ("FIDDLES", "FADDLES"),
-            True,
-            False,
-            id="skip_empty_values",
-        ),
-        pytest.param(
-            {"FIDDLES": "1", "FADDLES": "arr"},
-            ("", "FIDDLES", "FADDLES"),
-            False,
-            True,
-            id="skip_empty_keys",
-        ),
-    ],
-)
-def test_cfg_getbool(
-    env: dict[str, str],
-    keys: tuple[str, ...],
-    default: bool,
-    expected: str | None,
-):
-    assert cfg.getbool(*keys, default=default, env=env) == expected
-
-
-@pytest.mark.parametrize(
-    ("env", "keys", "default", "expected"),
-    [
-        pytest.param(
-            {"FURBY_NAMES": "George, Ringo"},
-            ("FURBY_NAMES",),
-            ("Walrus", "EGGMAN"),
-            ["George", "Ringo"],
-            id="simple",
-        ),
-        pytest.param(
-            {"FURBY_NAMES": "George, Ringo"},
-            ("FORBY_NAMES",),
-            ("Walrus", "EGGMAN"),
-            ["Walrus", "EGGMAN"],
-            id="missing",
-        ),
-        pytest.param(
-            {"ELEPHANTS": "Nacho Jack Cheddar", "FURBY_NAMES": "George, Ringo"},
-            ("FORBY_NAMES", "ELEPHANTS"),
-            ("Walrus", "EGGMAN"),
-            ["Nacho", "Jack", "Cheddar"],
-            id="first_wins",
-        ),
-        pytest.param(
-            {"ELEPHANTS": "", "FURBY_NAMES": "George Ringo"},
-            ("ELEPHANTS", "FURBY_NAMES"),
-            ("Walrus", "EGGMAN"),
-            ["George", "Ringo"],
-            id="skip_empty_values",
-        ),
-        pytest.param(
-            {"ELEPHANTS": "Nachorb,Jack%20Cheddar", "FURBY_NAMES": "George, Ringo"},
-            ("", "ELEPHANTS", "FURBY_NAMES"),
-            ("Walrus", "EGGMAN"),
-            ["Nachorb", "Jack%20Cheddar"],
-            id="skip_empty_keys",
-        ),
-    ],
-)
-def test_cfg_getlist(
-    env: dict[str, str],
-    keys: tuple[str, ...],
-    default: tuple[str, ...],
-    expected: list[str],
-):
-    assert cfg.getlist(*keys, default=default, env=env) == expected
-
-
-@pytest.mark.parametrize(
-    ("env", "keys", "default", "expected"),
-    [
-        pytest.param(
-            {"CEREAL_MASCOTS": "corn_flakes:Rooster, fruit_balls:Rabbit"},
-            ("CEREAL_MASCOTS",),
-            {},
-            {"corn_flakes": "Rooster", "fruit_balls": "Rabbit"},
-            id="simple",
-        ),
-        pytest.param(
-            {"CEREAL_MASCOTS": "corn_flakes:Rooster, fruit_balls:Rabbit"},
-            ("FURNISHINGS",),
-            {"frog": "true", "egg": "1"},
-            {"frog": "true", "egg": "1"},
-            id="missing",
-        ),
-        pytest.param(
-            {
-                "WEASELS": "Nacho:Jack Cheddar:fella",
-                "CEREAL_MASCOTS": "corn_flakes:Rooster, fruit_balls:Rabbit",
-            },
-            ("FORBY_NAMES", "WEASELS"),
-            {"frog": "true", "egg": "1"},
-            {"Nacho": "Jack", "Cheddar": "fella"},
-            id="first_wins",
-        ),
-        pytest.param(
-            {
-                "WEASELS": "",
-                "CEREAL_MASCOTS": "corn_flakes:Rooster, fruit_balls:Rabbit",
-            },
-            ("WEASELS", "CEREAL_MASCOTS"),
-            {"frog": "true", "egg": "1"},
-            {"corn_flakes": "Rooster", "fruit_balls": "Rabbit"},
-            id="skip_empty_values",
-        ),
-        pytest.param(
-            {
-                "WEASELS": "Nachorb:CHEZ,Jack%20Cheddar:boi",
-                "CEREAL_MASCOTS": "corn_flakes:Rooster, fruit_balls:Rabbit",
-            },
-            ("", "WEASELS", "CEREAL_MASCOTS"),
-            {"frog": "true", "egg": "1"},
-            {"Nachorb": "CHEZ", "Jack%20Cheddar": "boi"},
-            id="skip_empty_keys",
-        ),
-    ],
-)
-def test_cfg_getdict(
-    env: dict[str, str],
-    keys: tuple[str, ...],
-    default: dict[str, str],
-    expected: dict[str, str],
-):
-    assert cfg.getdict(*keys, default=default, env=env) == expected
